// api/resources/assets/client/src/app/models/ReviewLink.ts
export class ReviewLink {
  id: number;
  logo: string;
  description: string;
  name: string;
  url_slug: string;
  campaign_id: any;
  auto_approve: number;
  min_rating: number;
  negative_info_review_msg_1: string;
  negative_info_review_msg_2: string;
  positive_review_msg: string;
  created_by: number;
  deleted_at: any;
  created_at: any;
  updated_at: any;
  campaign?: any;
}